requirements.txt
streamlit
langchain
sentence-transformers
scikit-learn
torch
faiss-cpu
pypdf
python-dotenv
